# src/openapi.yaml
openapi: 3.1.0
info:
  title: Volvix POS API
  description: |
    API serverless de Volvix POS conectada a Supabase. Multi-tenant, multi-rol (cajero, admin, owner, superadmin).
    Autenticación basada en JWT (Bearer). Todas las rutas bajo `/api/*` salvo `/api/login`, `/api/logout`,
    `/api/health`, `/api/health/deep`, `/api/errors/log`, y los endpoints de password-reset requieren `Authorization: Bearer <jwt>`.
  version: 7.3.0-r14
  contact:
    name: Volvix POS
    url: https://volvix-pos.vercel.app
  license:
    name: Proprietary
servers:
  - url: https://volvix-pos.vercel.app
    description: Producción
  - url: http://localhost:3000
    description: Desarrollo local
security:
  - BearerAuth: []
...
paths:
...
  /api/products:
    get:
      tags: [Products]
      summary: Listar productos del tenant
      parameters:
        - in: query
          name: tenant_id
          schema: { type: string }
      responses:
        '200': { description: OK, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Product' } } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/ServerError' }
    post:
      tags: [Products]
      summary: Crear producto
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProductInput' }
      responses:
        '200': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/Product' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/ServerError' }
...
  /api/sales:
    get:
      tags: [Sales]
      summary: Listar ventas
      parameters:
        - in: query
          name: user_id
          schema: { type: string, format: uuid }
      responses:
        '200': { description: OK, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Sale' } } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/ServerError' }
    post:
      tags: [Sales]
      summary: Registrar venta (envía recibo por email si customer.email)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SaleInput' }
      responses:
        '200': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/Sale' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/ServerError' }
  /api/customers:
    get:
      tags: [Customers]
      summary: Listar clientes
      responses:
        '200': { description: OK, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Customer' } } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [Customers]
      summary: Crear cliente
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/CustomerInput' } } }
      responses:
        '200': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/Customer' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/ServerError' }
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    BadRequest:
      description: Petición inválida
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Unauthorized:
      description: No autenticado o JWT inválido
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
    ServerError:
      description: Error interno
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
...
    Product:
      type: object
      properties:
        id: { type: string, format: uuid }
        code: { type: string }
        name: { type: string }
        category: { type: string }
        price: { type: number }
        cost: { type: number }
        stock: { type: number }
        icon: { type: string }
        tenant_id: { type: string }
    ProductInput:
      type: object
      required: [name, price]
      properties:
        code: { type: string }
        name: { type: string }
        category: { type: string }
        cost: { type: number }
        price: { type: number }
        stock: { type: number }
        icon: { type: string }
        pos_user_id: { type: string, format: uuid }
    SaleItem:
      type: object
      properties:
        product_id: { type: string, format: uuid }
        code: { type: string }
        name: { type: string }
        qty: { type: number }
        price: { type: number }
        subtotal: { type: number }
    Sale:
      type: object
      properties:
        id: { type: string, format: uuid }
        pos_user_id: { type: string, format: uuid }
        total: { type: number }
        payment_method: { type: string, enum: [efectivo, tarjeta, transferencia, credito, mixto] }
        items: { type: array, items: { $ref: '#/components/schemas/SaleItem' } }
        created_at: { type: string, format: date-time }
    SaleInput:
      type: object
      required: [total]
      properties:
        total: { type: number }
        payment_method: { type: string }
        items: { type: array, items: { $ref: '#/components/schemas/SaleItem' } }
        customer:
          type: object
          properties:
            email: { type: string, format: email }
        customer_email: { type: string, format: email }
    Customer:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        email: { type: string, format: email, nullable: true }
        phone: { type: string, nullable: true }
        address: { type: string, nullable: true }
        credit_limit: { type: number }
        credit_balance: { type: number }
        points: { type: integer }
        loyalty_points: { type: integer }
        active: { type: boolean }
        user_id: { type: string, format: uuid }
        created_at: { type: string, format: date-time }
    CustomerInput:
      type: object
      required: [name]
      properties:
        name: { type: string }
        email: { type: string, format: email }
        phone: { type: string }
        address: { type: string }
        credit_limit: { type: number }
        credit_balance: { type: number }
        points: { type: integer }
        loyalty_points: { type: integer }
...